# openapi/pi-workflow.yaml
openapi: 3.1.0
info:
  title: PI Automation API
  description: API for automating personal injury case management workflows
  version: 0.1.0
paths:
...
  /api/cases:
    get:
      tags:
        - pi-workflow
      summary: List Cases
      operationId: list_cases_api_cases_get
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/cases/{caseId}:
    get:
      tags:
        - pi-workflow
      summary: Get Case Detail
      operationId: get_case_detail_api_cases__caseId__get
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            title: Caseid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/cases/{caseId}/advance:
    post:
      tags:
        - pi-workflow
      summary: Advance Case Stage
      operationId: advance_case_stage_api_cases__caseId__advance_post
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            title: Caseid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tasks:
    get:
      tags:
        - pi-workflow
      summary: List Tasks
      operationId: list_tasks_api_tasks_get
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      tags:
        - pi-workflow
      summary: Create Task
      operationId: create_task_api_tasks_post
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/tasks/{taskId}:
    patch:
      tags:
        - pi-workflow
      summary: Update Task
      operationId: update_task_api_tasks__taskId__patch
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            title: Taskid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - pi-workflow
      summary: Delete Task
      operationId: delete_task_api_tasks__taskId__delete
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            title: Taskid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tasks/bulk-complete:
    post:
      tags:
        - pi-workflow
      summary: Mark Many Tasks Done
      operationId: mark_many_tasks_done_api_tasks_bulk_complete_post
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/documents:
    get:
      tags:
        - pi-workflow
      summary: List Documents
      operationId: list_documents_api_documents_get
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/documents/{docId}/send:
    post:
      tags:
        - pi-workflow
      summary: Send Document
      operationId: send_document_api_documents__docId__send_post
      security:
        - BearerAuth: []
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: string
            title: Docid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT